requirements.txt
streamlit
transformers
torch
Pillow
huggingface_hub
gradio_client
openai==0.28.0
